LPS-66938 Move layout-admin-web data handlers to the proper package
